adapt tests for Carp 1.25 (RT #75383) and newer perls (RT #100500)
Carp 1.25 added an extra dot to the end of error messages.

Additionally, starting with perl 5.21.x a "use strict" violation has
additional diagnostics in the error message.

This change is adapting the tests so that the test suite should run
successfully with newer and older Carp, and with older and newer
perl versions.