Slay/Makefile.pm - strip off "Execution ... aborted" msg from eval errors
In https://github.com/Perl/perl5/pull/20357 the Perl dev team fixed some
inconsistencies in how eval behaved with regard to compile errors. Prior
that PR the exact outcome of a compile error varied with the type and
number of errors that were encountered, sometimes calling $SIG{__DIE__},
sometimes not, sometimes outputing "Execution ... aborted" messages,
sometimes not. With the merge of that PR the behavior is now consistent,
and in particular for this module the "Execution ... aborted" message is
now always output when there is a compile error.

This patch strips this message off the error before rethrowing the
exception with _croak().

This patch also includes some minor changes to the code to make it
easier to debug the code being eval()ed which I left commented out. I
figured if it was useful to debug this issue then it might useful again
in the future.